LayoutTests/platform/ipad/fast/forms/select-with-title-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS popoverTitle is document.getElementById('select').title
